# internal/gateway/handler/http/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GitHub Repository Gateway API",
    "version": "1.0.0",
    "description": "REST API gateway for fetching GitHub repository metadata through the collector gRPC service."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
    "/api/v1/repositories/{owner}/{repo}": {
      "get": {
        "summary": "Get repository information",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Repository info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Repository"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Repository not found"
          },
          "502": {
            "description": "Collector unavailable"
          },
          "504": {
            "description": "Upstream timeout"
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "Get OpenAPI specification",
        "responses": {
          "200": {
            "description": "OpenAPI JSON"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Repository": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "stars": {
            "type": "integer",
            "format": "int64"
          },
          "forks": {
            "type": "integer",
            "format": "int64"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "description",
          "stars",
          "forks",
          "created_at"
        ]
      }
    }
  }
}